requirements.txt
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
defusedxml==0.5.0
Django==2.1.4
django-braces==1.13.0
django-cors-headers==2.4.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.9.0
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.12.0
idna==2.8
oauthlib==2.1.0
Pillow==5.3.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyJWT==1.7.1
python-dotenv==0.10.1
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
rest-social-auth==1.5.0
rsa==4.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==1.7.0
socials==0.2.0
uritemplate==3.0.0
urllib3==1.24.1
